pyproject.toml
[project]
name = "talky-workspace"
version = "0.1.0"
description = "Talky voice AI workspace"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "loguru>=0.7.3",
]

[tool.uv.workspace]
members = ["server", "mcp-server"]

[dependency-groups]
dev = [
    "ruff>=0.12.11,<1",
    "pyright>=1.1.404,<2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["I"]

[project.scripts]
talky = "talky_cli:main"

[tool.setuptools]
py-modules = ["talky_cli", "shared"]

[tool.setuptools.packages.find]
include = ["talky_cli*", "shared*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
